app/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
psycopg2
requests
python-dotenv
joblib
scikit-learn
